refactor: rename comp
