+++ apps/web/src/views/Swap/MMLinkPools/components/MMCommitButton.tsx
@@ -1,5 +1,5 @@
 import { useTranslation } from '@pancakeswap/localization'
-import { Currency, CurrencyAmount, Token, TradeType } from '@pancakeswap/sdk'
+import { Currency, CurrencyAmount, TradeType } from '@pancakeswap/sdk'
 import { SmartRouterTrade } from '@pancakeswap/smart-router/evm'
 import { Button, useModal } from '@pancakeswap/uikit'
 import { useExpertMode } from '@pancakeswap/utils/user'
@@ -7,31 +7,33 @@ import { CommitButton } from 'components/CommitButton'
 import ConnectWalletButton from 'components/ConnectWalletButton'
 import SettingsModal, { withCustomOnDismiss } from 'components/Menu/GlobalSettings/SettingsModal'
 import { SettingsMode } from 'components/Menu/GlobalSettings/types'
+import { useActiveChainId } from 'hooks/useActiveChainId'
+import { ApprovalState } from 'hooks/useApproveCallback'
 import { WrapType } from 'hooks/useWrapCallback'
-import { useCallback, useEffect, useMemo, useState } from 'react'
+import { useCallback, useEffect, useState } from 'react'
 import { Field } from 'state/swap/actions'
 import { logGTMClickSwapEvent } from 'utils/customGTMEventTracking'
-import { Address } from 'viem'
 import { parseMMError } from 'views/Swap/MMLinkPools/utils/exchange'
 import { ConfirmSwapModal } from 'views/Swap/V3Swap/containers/ConfirmSwapModal'
-import { MMTradeInfo } from '../hooks'
-import { useMMConfirmModalState } from '../hooks/useMMConfirmModalState'
+import { useConfirmModalState } from 'views/Swap/V3Swap/hooks/useConfirmModalState'
+import { SendTransactionResult } from 'wagmi/actions'
 import { useSwapCallArguments } from '../hooks/useSwapCallArguments'
+import { useSwapCallback } from '../hooks/useSwapCallback'
 import { MMRfqTrade } from '../types'
 
 const SettingsModalWithCustomDismiss = withCustomOnDismiss(SettingsModal)
 
 interface SwapCommitButtonPropsType {
-  mmTradeInfo: MMTradeInfo
   swapIsUnsupported: boolean
-  account: Address | undefined
+  account: string | undefined
   showWrap: boolean
-  wrapInputError?: string
+  wrapInputError: string | undefined
   onWrap?: () => Promise<void>
   wrapType: WrapType
-  // approval: ApprovalState
-  // allowance: Allowance
-  // approvalSubmitted: boolean
+  approval: ApprovalState
+  approveCallback: () => Promise<SendTransactionResult | undefined>
+  revokeCallback: () => Promise<SendTransactionResult | undefined>
+  approvalSubmitted: boolean
   currencies: {
     INPUT?: Currency
     OUTPUT?: Currency
@@ -46,58 +48,100 @@ interface SwapCommitButtonPropsType {
   recipient: string | null
   onUserInput: (field: Field, typedValue: string) => void
   mmQuoteExpiryRemainingSec?: number | null
-  // isPendingError: boolean
-  // currentAllowance?: CurrencyAmount<Currency>
+  isPendingError: boolean
+  currentAllowance: CurrencyAmount<Currency> | undefined
 }
 
 export function MMSwapCommitButton({
-  mmTradeInfo,
   swapIsUnsupported,
   account,
   showWrap,
   wrapInputError,
   onWrap,
   wrapType,
+  approval,
+  approveCallback,
+  revokeCallback,
+  approvalSubmitted,
   rfqTrade,
   swapInputError,
   currencyBalances,
   recipient,
   onUserInput,
-}: // isPendingError,
-SwapCommitButtonPropsType) {
+  isPendingError,
+  currentAllowance,
+}: SwapCommitButtonPropsType) {
+  const { chainId } = useActiveChainId()
+
   const [isExpertMode] = useExpertMode()
 
   const { t } = useTranslation()
-  const [tradeToConfirm, setTradeToConfirm] = useState<SmartRouterTrade<TradeType> | undefined>(undefined)
-  const amountToApprove = useMemo(() => {
-    return mmTradeInfo?.slippageAdjustedAmounts[Field.INPUT]
-  }, [mmTradeInfo])
-  const mmSpender = useMemo(() => {
-    return mmTradeInfo?.routerAddress as Address | undefined
-  }, [mmTradeInfo])
-  const swapCalls = useSwapCallArguments(rfqTrade.trade, rfqTrade.rfq ?? undefined, recipient ?? undefined)
-  const { resetState, txHash, confirmState, confirmSteps, callToAction, errorMessage } = useMMConfirmModalState(
+  // the callback to execute the swap
+
+  // @ts-ignore
+  const swapCalls = useSwapCallArguments(rfqTrade.trade, rfqTrade.rfq, recipient)
+
+  const { callback: swapCallback, error: swapCallbackError } = useSwapCallback(
     rfqTrade.trade,
+    recipient,
+    // @ts-ignore
     swapCalls,
-    (recipient as Address) ?? null,
-    amountToApprove?.currency.isToken ? (amountToApprove as CurrencyAmount<Token>) : undefined,
-    mmSpender,
   )
+  const [{ tradeToConfirm, swapErrorMessage, attemptingTxn, txHash }, setSwapState] = useState<{
+    tradeToConfirm?: SmartRouterTrade<TradeType>
+    attemptingTxn: boolean
+    swapErrorMessage: string | undefined
+    txHash: string | undefined
+  }>({
+    tradeToConfirm: undefined,
+    attemptingTxn: false,
+    swapErrorMessage: undefined,
+    txHash: undefined,
+  })
 
   // Handlers
+  const handleSwap = useCallback(async () => {
+    if (!swapCallback) {
+      return undefined
+    }
+    setSwapState({ attemptingTxn: true, tradeToConfirm, swapErrorMessage: undefined, txHash: undefined })
+    return swapCallback()
+      .then((result) => {
+        setSwapState({ attemptingTxn: false, tradeToConfirm, swapErrorMessage: undefined, txHash: result.hash })
+      })
+      .catch((error) => {
+        console.error('handleSwap error', error)
+
+        setSwapState({
+          attemptingTxn: false,
+          tradeToConfirm,
+          swapErrorMessage: error.message,
+          txHash: undefined,
+        })
+      })
+  }, [swapCallback, tradeToConfirm, setSwapState])
 
   const handleAcceptChanges = useCallback(() => {
-    setTradeToConfirm(rfqTrade.trade)
-    resetState()
-  }, [rfqTrade.trade, resetState])
+    setSwapState({ tradeToConfirm: rfqTrade.trade, swapErrorMessage, txHash, attemptingTxn })
+  }, [attemptingTxn, swapErrorMessage, rfqTrade, txHash, setSwapState])
 
   const handleConfirmDismiss = useCallback(() => {
-    resetState()
+    setSwapState({ tradeToConfirm, attemptingTxn, swapErrorMessage, txHash })
     // if there was a tx hash, we want to clear the input
     if (txHash) {
       onUserInput(Field.INPUT, '')
     }
-  }, [resetState, txHash, onUserInput])
+  }, [attemptingTxn, onUserInput, swapErrorMessage, tradeToConfirm, txHash, setSwapState])
+
+  // End Handlers
+
+  // show approve flow when: no error on inputs, not approved or pending, or approved in current session
+  // never show if price impact is above threshold in non expert mode
+  const showApproveFlow =
+    !swapInputError &&
+    (approval === ApprovalState.NOT_APPROVED ||
+      approval === ApprovalState.PENDING ||
+      (approvalSubmitted && approval === ApprovalState.APPROVED))
 
   // Modals
   const [indirectlyOpenConfirmModalState, setIndirectlyOpenConfirmModalState] = useState(false)
@@ -109,18 +153,35 @@ SwapCommitButtonPropsType) {
     />,
   )
 
+  const { confirmModalState, pendingModalSteps, startSwapFlow, resetSwapFlow } = useConfirmModalState({
+    txHash,
+    chainId,
+    approval,
+    approvalToken: rfqTrade?.trade?.inputAmount.currency,
+    isPendingError,
+    isExpertMode,
+    currentAllowance,
+    approveCallback,
+    revokeCallback,
+    onConfirm: handleSwap,
+  })
+
   const [onPresentConfirmModal] = useModal(
     <ConfirmSwapModal
       isMM
       trade={rfqTrade.trade} // show the info while refresh RFQ
-      originalTrade={tradeToConfirm}
       txHash={txHash}
-      confirmModalState={confirmState}
-      pendingModalSteps={confirmSteps ?? []}
-      onConfirm={callToAction}
+      approval={approval}
+      attemptingTxn={attemptingTxn}
+      confirmModalState={confirmModalState}
+      pendingModalSteps={pendingModalSteps}
+      startSwapFlow={startSwapFlow}
+      originalTrade={tradeToConfirm}
+      showApproveFlow={showApproveFlow}
       currencyBalances={currencyBalances}
       isRFQReady={Boolean(rfqTrade.rfq) && !rfqTrade.isLoading}
-      swapErrorMessage={errorMessage || (!rfqTrade.trade ? t('Unable request a quote') : undefined)}
+      currentAllowance={currentAllowance}
+      swapErrorMessage={swapErrorMessage || (!rfqTrade.trade && t('Unable request a quote'))}
       onAcceptChanges={handleAcceptChanges}
       customOnDismiss={handleConfirmDismiss}
       openSettingModal={onPresentSettingsModal}
@@ -132,24 +193,31 @@ SwapCommitButtonPropsType) {
   // End Modals
 
   const onSwapHandler = useCallback(() => {
-    setTradeToConfirm(rfqTrade.trade)
-    resetState()
-
+    setSwapState({
+      tradeToConfirm: rfqTrade.trade,
+      attemptingTxn: false,
+      swapErrorMessage: undefined,
+      txHash: undefined,
+    })
+    resetSwapFlow()
     if (isExpertMode) {
-      callToAction()
+      startSwapFlow()
     }
     onPresentConfirmModal()
     logGTMClickSwapEvent()
-  }, [rfqTrade.trade, resetState, isExpertMode, onPresentConfirmModal, callToAction])
+  }, [rfqTrade, onPresentConfirmModal, isExpertMode, startSwapFlow, resetSwapFlow])
 
   // useEffect
   useEffect(() => {
     if (indirectlyOpenConfirmModalState) {
       setIndirectlyOpenConfirmModalState(false)
-      resetState()
+      setSwapState((state) => ({
+        ...state,
+        swapErrorMessage: undefined,
+      }))
       onPresentConfirmModal()
     }
-  }, [indirectlyOpenConfirmModalState, onPresentConfirmModal, resetState])
+  }, [indirectlyOpenConfirmModalState, onPresentConfirmModal, setSwapState])
 
   if (swapIsUnsupported) {
     return (
@@ -183,7 +251,7 @@ SwapCommitButtonPropsType) {
       width="100%"
       id="swap-button"
       variant="primary"
-      disabled={!rfqTrade.rfq || !isValid || !!errorMessage}
+      disabled={!rfqTrade.rfq || !isValid || !!swapCallbackError}
       onClick={onSwapHandler}
       data-dd-action-name="Swap mm commit button"
     >
