+++ apps/web/src/views/Swap/V3Swap/containers/ApproveStepFlow.tsx
@@ -1,9 +1,8 @@
 import { useTranslation } from '@pancakeswap/localization'
 import { Box, Flex, LinkExternal, Text, useTooltip } from '@pancakeswap/uikit'
-import { ConfirmModalState } from '@pancakeswap/widgets-internal'
 import { useMemo } from 'react'
 import { styled } from 'styled-components'
-import { PendingConfirmModalState } from '../types'
+import { ConfirmModalStateV1, PendingConfirmModalStateV1 } from '../types'
 
 const StyledLinkExternal = styled(LinkExternal)`
   &:hover {
@@ -27,27 +26,8 @@ const Step = styled('div')<{ active: boolean; width: string }>`
 `
 
 interface ApproveStepFlowProps {
-  confirmModalState: ConfirmModalState
-  pendingModalSteps: PendingConfirmModalState[]
-}
-
-const ApprovalSteps = ({ pendingModalSteps, confirmModalState }: ApproveStepFlowProps) => {
-  const stepWidth = useMemo(() => `${100 / pendingModalSteps.length}%`, [pendingModalSteps])
-  return (
-    <>
-      {pendingModalSteps.map((pendingStep: PendingConfirmModalState) => {
-        return (
-          <Step
-            key={pendingStep}
-            active={
-              Object.values(ConfirmModalState)[confirmModalState] === Object.values(ConfirmModalState)[pendingStep]
-            }
-            width={stepWidth}
-          />
-        )
-      })}
-    </>
-  )
+  confirmModalState: ConfirmModalStateV1
+  pendingModalSteps: PendingConfirmModalStateV1[]
 }
 
 export const ApproveStepFlow: React.FC<React.PropsWithChildren<ApproveStepFlowProps>> = ({
@@ -64,6 +44,7 @@ export const ApproveStepFlow: React.FC<React.PropsWithChildren<ApproveStepFlowPr
     { placement: 'top' },
   )
 
+  const stepWidth = useMemo(() => `${100 / pendingModalSteps.length}%`, [pendingModalSteps])
   const hideStepIndicators = useMemo(() => pendingModalSteps.length === 1, [pendingModalSteps])
 
   return (
@@ -74,9 +55,14 @@ export const ApproveStepFlow: React.FC<React.PropsWithChildren<ApproveStepFlowPr
       {!hideStepIndicators && (
         <>
           <StepsContainer>
-            <ApprovalSteps pendingModalSteps={pendingModalSteps} confirmModalState={confirmModalState} />
+            {pendingModalSteps.length !== 3 && (
+              <Step active={confirmModalState === ConfirmModalStateV1.RESETTING_APPROVAL} width={stepWidth} />
+            )}
+            <Step active={confirmModalState === ConfirmModalStateV1.APPROVING_TOKEN} width={stepWidth} />
+            <Step active={confirmModalState === ConfirmModalStateV1.APPROVE_PENDING} width={stepWidth} />
+            <Step active={confirmModalState === ConfirmModalStateV1.PENDING_CONFIRMATION} width={stepWidth} />
           </StepsContainer>
-          {confirmModalState === ConfirmModalState.RESETTING_APPROVAL && (
+          {confirmModalState === ConfirmModalStateV1.RESETTING_APPROVAL && (
             <StyledLinkExternal
               external
               margin="16px auto auto auto"
@@ -85,7 +71,7 @@ export const ApproveStepFlow: React.FC<React.PropsWithChildren<ApproveStepFlowPr
               <Text color="primary">{t('Why resetting approval')}</Text>
             </StyledLinkExternal>
           )}
-          {confirmModalState === ConfirmModalState.APPROVING_TOKEN && (
+          {confirmModalState === ConfirmModalStateV1.APPROVING_TOKEN && (
             <StyledLinkExternal
               external
               margin="16px auto auto auto"
