+++ apps/web/src/views/Swap/V3Swap/containers/ConfirmSwapModal.tsx
@@ -1,118 +1,113 @@
+import { ChainId } from '@pancakeswap/chains'
+import { Currency, CurrencyAmount, Token, TradeType } from '@pancakeswap/sdk'
+import { memo, useCallback, useMemo } from 'react'
+
 import { useTranslation } from '@pancakeswap/localization'
-import { ChainId, Currency, CurrencyAmount, Token, TradeType } from '@pancakeswap/sdk'
-import { SmartRouterTrade } from '@pancakeswap/smart-router'
+import { SmartRouterTrade } from '@pancakeswap/smart-router/evm'
 import { WrappedTokenInfo } from '@pancakeswap/token-lists'
 import { Box, BscScanIcon, Flex, InjectedModalProps, Link } from '@pancakeswap/uikit'
 import { formatAmount } from '@pancakeswap/utils/formatFractions'
 import truncateHash from '@pancakeswap/utils/truncateHash'
-import { useUserSlippage } from '@pancakeswap/utils/user'
 import {
-  ApproveModalContent,
-  ConfirmModalState,
-  SwapPendingModalContent,
-  SwapTransactionReceiptModalContent,
+  ApproveModalContentV1,
+  SwapPendingModalContentV1,
+  SwapTransactionReceiptModalContentV1,
 } from '@pancakeswap/widgets-internal'
-import AddToWalletButton, { AddToWalletTextOptions } from 'components/AddToWallet/AddToWalletButton'
-import { useActiveChainId } from 'hooks/useActiveChainId'
-import { useCallback, useMemo } from 'react'
-import { Field } from 'state/swap/actions'
-import { useSwapState } from 'state/swap/hooks'
 import { getBlockExploreLink, getBlockExploreName } from 'utils'
 import { wrappedCurrency } from 'utils/wrappedCurrency'
-import ConfirmSwapModalContainer from 'views/Swap/components/ConfirmSwapModalContainer'
-import { SwapTransactionErrorContent } from 'views/Swap/components/SwapTransactionErrorContent'
+
+import { useDebounce } from '@pancakeswap/hooks'
+import { useUserSlippage } from '@pancakeswap/utils/user'
+import AddToWalletButton, { AddToWalletTextOptions } from 'components/AddToWallet/AddToWalletButton'
+import { useActiveChainId } from 'hooks/useActiveChainId'
+import { ApprovalState } from 'hooks/useApproveCallback'
+import { Field } from 'state/swap/actions'
+import { useSwapState } from 'state/swap/hooks'
+import { ConfirmModalStateV1, PendingConfirmModalStateV1 } from '../types'
+
+import ConfirmSwapModalContainer from '../../components/ConfirmSwapModalContainer'
+import { SwapTransactionErrorContent } from '../../components/SwapTransactionErrorContent'
 import { TransactionConfirmSwapContent } from '../components'
-import { ConfirmAction } from '../hooks/useConfirmModalState'
-import { AllowedAllowanceState } from '../types'
+import { useWallchainStatus } from '../hooks/useWallchain'
 import { ApproveStepFlow } from './ApproveStepFlow'
 
-export const useApprovalPhaseStepTitles: ({ trade }: { trade: SmartRouterTrade<TradeType> | undefined }) => {
-  [step in AllowedAllowanceState]: string
-} = ({ trade }: { trade: SmartRouterTrade<TradeType> | undefined }) => {
-  const { t } = useTranslation()
-  return useMemo(() => {
-    return {
-      [ConfirmModalState.RESETTING_APPROVAL]: t('Reset approval on USDT.'),
-      [ConfirmModalState.APPROVING_TOKEN]: t('Approve %symbol%', {
-        symbol: trade ? trade.inputAmount.currency.symbol : '',
-      }),
-      [ConfirmModalState.PERMITTING]: t('Permit %symbol%', { symbol: trade ? trade.inputAmount.currency.symbol : '' }),
-    }
-  }, [t, trade])
-}
-
-type ConfirmSwapModalProps = InjectedModalProps & {
-  customOnDismiss?: () => void
-  onDismiss?: () => void
-  confirmModalState: ConfirmModalState
-  pendingModalSteps: ConfirmAction[]
+interface ConfirmSwapModalProps {
   isMM?: boolean
   isRFQReady?: boolean
   trade?: SmartRouterTrade<TradeType>
   originalTrade?: SmartRouterTrade<TradeType>
   currencyBalances: { [field in Field]?: CurrencyAmount<Currency> }
+  attemptingTxn: boolean
   txHash?: string
-  swapErrorMessage?: string
+  approval: ApprovalState
+  swapErrorMessage?: string | boolean
+  showApproveFlow: boolean
+  confirmModalState: ConfirmModalStateV1
+  startSwapFlow: () => void
+  pendingModalSteps: PendingConfirmModalStateV1[]
+  currentAllowance?: CurrencyAmount<Currency>
   onAcceptChanges: () => void
-  onConfirm: (setConfirmModalState?: () => void) => void
+  customOnDismiss?: () => void
   openSettingModal?: () => void
 }
 
-export const ConfirmSwapModal: React.FC<ConfirmSwapModalProps> = ({
+export const ConfirmSwapModal = memo<InjectedModalProps & ConfirmSwapModalProps>(function ConfirmSwapModalComp({
+  isMM,
+  trade,
+  txHash,
   confirmModalState,
+  startSwapFlow,
   pendingModalSteps,
-  customOnDismiss,
+  isRFQReady,
+  attemptingTxn,
+  originalTrade,
+  showApproveFlow,
+  currencyBalances,
   swapErrorMessage,
   onDismiss,
-  isMM,
-  isRFQReady,
-  trade,
-  originalTrade,
-  txHash,
-  currencyBalances,
-  openSettingModal,
   onAcceptChanges,
-  onConfirm,
-}) => {
-  const { t } = useTranslation()
+  customOnDismiss,
+  openSettingModal,
+}) {
   const { chainId } = useActiveChainId()
+  const { t } = useTranslation()
   const [allowedSlippage] = useUserSlippage()
   const { recipient } = useSwapState()
-  const loadingAnimationVisible = useMemo(() => {
-    return [
-      ConfirmModalState.RESETTING_APPROVAL,
-      ConfirmModalState.APPROVING_TOKEN,
-      ConfirmModalState.PERMITTING,
-    ].includes(confirmModalState)
-  }, [confirmModalState])
-  const hasError = useMemo(() => swapErrorMessage !== undefined, [swapErrorMessage])
-  const stepsVisible = useMemo(() => {
-    if (swapErrorMessage) return false
-    if (confirmModalState === ConfirmModalState.REVIEWING || confirmModalState === ConfirmModalState.COMPLETED)
-      return false
-    if (confirmModalState === ConfirmModalState.PENDING_CONFIRMATION && txHash) return false
-    return pendingModalSteps.length > 0 && pendingModalSteps.some((step) => step.showIndicator)
-  }, [confirmModalState, pendingModalSteps, swapErrorMessage, txHash])
+  const [wallchainStatus] = useWallchainStatus()
+  const isBonus = useDebounce(wallchainStatus === 'found', 500)
 
-  const stepContents = useApprovalPhaseStepTitles({ trade })
-  const token: Token | undefined = useMemo(
-    () => wrappedCurrency(trade?.outputAmount?.currency, chainId),
-    [chainId, trade?.outputAmount?.currency],
-  )
+  const token: Token | undefined = wrappedCurrency(trade?.outputAmount?.currency, chainId)
 
   const handleDismiss = useCallback(() => {
-    if (typeof customOnDismiss === 'function') {
-      customOnDismiss()
+    if (customOnDismiss) {
+      customOnDismiss?.()
     }
-
     onDismiss?.()
   }, [customOnDismiss, onDismiss])
 
-  const modalContent = useMemo(() => {
+  const topModal = useMemo(() => {
     const currencyA = currencyBalances.INPUT?.currency ?? trade?.inputAmount?.currency
     const currencyB = currencyBalances.OUTPUT?.currency ?? trade?.outputAmount?.currency
     const amountA = formatAmount(trade?.inputAmount, 6) ?? ''
     const amountB = formatAmount(trade?.outputAmount, 6) ?? ''
+
+    if (confirmModalState === ConfirmModalStateV1.RESETTING_APPROVAL) {
+      return <ApproveModalContentV1 title={t('Reset Approval on USDT')} isMM={isMM} isBonus={isBonus} />
+    }
+
+    if (
+      showApproveFlow &&
+      (confirmModalState === ConfirmModalStateV1.APPROVING_TOKEN ||
+        confirmModalState === ConfirmModalStateV1.APPROVE_PENDING)
+    ) {
+      return (
+        <ApproveModalContentV1
+          title={t('Enable spending %symbol%', { symbol: `${trade?.inputAmount?.currency?.symbol}` })}
+          isMM={isMM}
+          isBonus={isBonus}
+        />
+      )
+    }
 
     if (swapErrorMessage) {
       return (
@@ -125,67 +120,28 @@ export const ConfirmSwapModal: React.FC<ConfirmSwapModalProps> = ({
         </Flex>
       )
     }
-    if (
-      confirmModalState === ConfirmModalState.APPROVING_TOKEN ||
-      confirmModalState === ConfirmModalState.PERMITTING ||
-      confirmModalState === ConfirmModalState.RESETTING_APPROVAL
-    ) {
-      return (
-        <ApproveModalContent
-          title={stepContents}
-          isMM={isMM}
-          // TODO
-          isBonus={false}
-          currencyA={currencyA as Currency}
-          asBadge
-          currentStep={confirmModalState}
-          approvalModalSteps={pendingModalSteps.map((step) => step.step) as any}
-        />
-      )
-    }
 
-    if (confirmModalState === ConfirmModalState.PENDING_CONFIRMATION) {
+    if (attemptingTxn) {
       return (
-        <SwapPendingModalContent
-          title={txHash ? t('Transaction Submitted') : t('Confirm Swap')}
+        <SwapPendingModalContentV1
+          title={t('Confirm Swap')}
           currencyA={currencyA}
           currencyB={currencyB}
           amountA={amountA}
           amountB={amountB}
-          currentStep={confirmModalState}
-        >
-          {txHash ? (
-            <AddToWalletButton
-              mt="39px"
-              height="auto"
-              variant="tertiary"
-              width="fit-content"
-              padding="6.5px 20px"
-              marginTextBetweenLogo="6px"
-              textOptions={AddToWalletTextOptions.TEXT_WITH_ASSET}
-              tokenAddress={token?.address}
-              tokenSymbol={currencyB?.symbol}
-              tokenDecimals={token?.decimals}
-              tokenLogo={token instanceof WrappedTokenInfo ? (token as WrappedTokenInfo)?.logoURI : undefined}
-            />
-          ) : null}
-        </SwapPendingModalContent>
+        />
       )
     }
 
-    if (confirmModalState === ConfirmModalState.COMPLETED && txHash) {
+    if (confirmModalState === ConfirmModalStateV1.PENDING_CONFIRMATION) {
       return (
-        <SwapTransactionReceiptModalContent
-          explorerLink={
-            chainId ? (
-              <Link external small href={getBlockExploreLink(txHash, 'transaction', chainId)}>
-                {t('View on %site%', { site: getBlockExploreName(chainId) })}: {truncateHash(txHash, 8, 0)}
-                {chainId === ChainId.BSC && <BscScanIcon color="primary" ml="4px" />}
-              </Link>
-            ) : (
-              <></>
-            )
-          }
+        <SwapPendingModalContentV1
+          showIcon
+          title={t('Transaction Submitted')}
+          currencyA={currencyA}
+          currencyB={currencyB}
+          amountA={amountA}
+          amountB={amountB}
         >
           <AddToWalletButton
             mt="39px"
@@ -198,66 +154,99 @@ export const ConfirmSwapModal: React.FC<ConfirmSwapModalProps> = ({
             tokenAddress={token?.address}
             tokenSymbol={currencyB?.symbol}
             tokenDecimals={token?.decimals}
-            tokenLogo={token instanceof WrappedTokenInfo ? (token as WrappedTokenInfo)?.logoURI : undefined}
+            tokenLogo={token instanceof WrappedTokenInfo ? token?.logoURI : undefined}
           />
-        </SwapTransactionReceiptModalContent>
+        </SwapPendingModalContentV1>
+      )
+    }
+
+    if (confirmModalState === ConfirmModalStateV1.COMPLETED && txHash) {
+      return (
+        <SwapTransactionReceiptModalContentV1>
+          {chainId && (
+            <Link external small href={getBlockExploreLink(txHash, 'transaction', chainId)}>
+              {t('View on %site%', { site: getBlockExploreName(chainId) })}: {truncateHash(txHash, 8, 0)}
+              {chainId === ChainId.BSC && <BscScanIcon color="primary" ml="4px" />}
+            </Link>
+          )}
+          <AddToWalletButton
+            mt="39px"
+            height="auto"
+            variant="tertiary"
+            width="fit-content"
+            padding="6.5px 20px"
+            marginTextBetweenLogo="6px"
+            textOptions={AddToWalletTextOptions.TEXT_WITH_ASSET}
+            tokenAddress={token?.address}
+            tokenSymbol={currencyB?.symbol}
+            tokenDecimals={token?.decimals}
+            tokenLogo={token instanceof WrappedTokenInfo ? token?.logoURI : undefined}
+          />
+        </SwapTransactionReceiptModalContentV1>
       )
     }
 
     return (
       <TransactionConfirmSwapContent
         isMM={isMM}
-        isRFQReady={isRFQReady}
         trade={trade}
         recipient={recipient}
+        isRFQReady={isRFQReady}
         originalTrade={originalTrade}
         allowedSlippage={allowedSlippage}
         currencyBalances={currencyBalances}
-        onConfirm={onConfirm}
+        onConfirm={startSwapFlow}
         onAcceptChanges={onAcceptChanges}
       />
     )
   }, [
-    allowedSlippage,
-    chainId,
-    confirmModalState,
-    currencyBalances,
-    handleDismiss,
     isMM,
-    isRFQReady,
-    onAcceptChanges,
-    onConfirm,
-    openSettingModal,
-    originalTrade,
-    pendingModalSteps,
-    recipient,
-    stepContents,
-    swapErrorMessage,
-    t,
-    token,
+    isBonus,
     trade,
     txHash,
+    isRFQReady,
+    originalTrade,
+    attemptingTxn,
+    currencyBalances,
+    showApproveFlow,
+    swapErrorMessage,
+    token,
+    chainId,
+    recipient,
+    allowedSlippage,
+    confirmModalState,
+    t,
+    handleDismiss,
+    startSwapFlow,
+    onAcceptChanges,
+    openSettingModal,
   ])
+
+  const isShowingLoadingAnimation = useMemo(
+    () =>
+      confirmModalState === ConfirmModalStateV1.RESETTING_APPROVAL ||
+      confirmModalState === ConfirmModalStateV1.APPROVING_TOKEN ||
+      confirmModalState === ConfirmModalStateV1.APPROVE_PENDING ||
+      attemptingTxn,
+    [confirmModalState, attemptingTxn],
+  )
 
   if (!chainId) return null
 
   return (
     <ConfirmSwapModalContainer
-      minHeight={hasError ? 'auto' : '415px'}
+      minHeight="415px"
       width={['100%', '100%', '100%', '367px']}
-      hideTitleAndBackground={confirmModalState !== ConfirmModalState.REVIEWING || hasError}
-      headerPadding={loadingAnimationVisible ? '12px 24px 0px 24px !important' : '12px 24px'}
-      bodyPadding={loadingAnimationVisible && !hasError ? '0 24px 24px 24px' : '24px'}
-      bodyTop={loadingAnimationVisible ? '-15px' : '0'}
+      hideTitleAndBackground={confirmModalState !== ConfirmModalStateV1.REVIEWING || Boolean(swapErrorMessage)}
+      headerPadding={isShowingLoadingAnimation ? '12px 24px 0px 24px !important' : '12px 24px'}
+      bodyPadding={isShowingLoadingAnimation ? '0 24px 24px 24px' : '24px'}
+      bodyTop={isShowingLoadingAnimation ? '-15px' : '0'}
       handleDismiss={handleDismiss}
     >
-      <Box>{modalContent}</Box>
-      {stepsVisible ? (
-        <ApproveStepFlow
-          confirmModalState={confirmModalState}
-          pendingModalSteps={pendingModalSteps.map((step) => step.step) as any}
-        />
-      ) : null}
+      <Box>{topModal}</Box>
+      {isShowingLoadingAnimation && !swapErrorMessage && (
+        <ApproveStepFlow confirmModalState={confirmModalState} pendingModalSteps={pendingModalSteps} />
+      )}
     </ConfirmSwapModalContainer>
   )
-}
+})
