+++ apps/web/src/views/Swap/V3Swap/containers/MMCommitButton.tsx
@@ -2,8 +2,9 @@ import { Currency } from '@pancakeswap/sdk'
 import { useExpertMode } from '@pancakeswap/utils/user'
 import { useCurrency } from 'hooks/Tokens'
 import { useIsTransactionUnsupported } from 'hooks/Trades'
+import { ApprovalState, useApproveCallback } from 'hooks/useApproveCallback'
 import useWrapCallback, { WrapType } from 'hooks/useWrapCallback'
-import { memo, useMemo } from 'react'
+import { memo, useEffect, useMemo, useState } from 'react'
 import { Field } from 'state/swap/actions'
 import { useSwapState } from 'state/swap/hooks'
 import { useSwapActionHandlers } from 'state/swap/useSwapActionHandlers'
@@ -18,14 +19,14 @@ function MMCommitButtonComp({ mmOrderBookTrade, mmRFQTrade, mmQuoteExpiryRemaini
     [Field.OUTPUT]: { currencyId: outputCurrencyId },
   } = useSwapState()
 
-  const inputCurrency = useCurrency(inputCurrencyId) ?? undefined
-  const outputCurrency = useCurrency(outputCurrencyId) ?? undefined
+  const inputCurrency = useCurrency(inputCurrencyId)
+  const outputCurrency = useCurrency(outputCurrencyId)
   const { address: account } = useAccount()
   const [isExpertMode] = useExpertMode()
   const currencies: { [field in Field]?: Currency } = useMemo(
     () => ({
-      [Field.INPUT]: inputCurrency,
-      [Field.OUTPUT]: outputCurrency,
+      [Field.INPUT]: inputCurrency ?? undefined,
+      [Field.OUTPUT]: outputCurrency ?? undefined,
     }),
     [inputCurrency, outputCurrency],
   )
@@ -37,15 +38,39 @@ function MMCommitButtonComp({ mmOrderBookTrade, mmRFQTrade, mmQuoteExpiryRemaini
     inputError: wrapInputError,
   } = useWrapCallback(inputCurrency, outputCurrency, typedValue)
   const showWrap = wrapType !== WrapType.NOT_APPLICABLE
+
+  const { approvalState, approveCallback, revokeCallback, currentAllowance, isPendingError } = useApproveCallback(
+    mmTradeInfo?.slippageAdjustedAmounts[Field.INPUT],
+    mmTradeInfo?.routerAddress,
+  )
+
+  // check if user has gone through approval process, used to show two step buttons, reset on token change
+  const [approvalSubmitted, setApprovalSubmitted] = useState<boolean>(false)
+
+  // Reset approval flow if input currency changed
+  useEffect(() => {
+    setApprovalSubmitted(false)
+  }, [inputCurrencyId])
+
+  // mark when a user has submitted an approval, reset onTokenSelection for input field
+  useEffect(() => {
+    if (approvalState === ApprovalState.PENDING) {
+      setApprovalSubmitted(true)
+    }
+  }, [approvalState, approvalSubmitted])
+
   const { onUserInput } = useSwapActionHandlers()
 
   return (
     <MMSwapCommitButton
-      mmTradeInfo={mmTradeInfo}
       showWrap={showWrap}
+      approval={approvalState}
       swapIsUnsupported={swapIsUnsupported}
       account={account}
+      approvalSubmitted={approvalSubmitted}
       onWrap={onWrap}
+      approveCallback={approveCallback}
+      revokeCallback={revokeCallback}
       currencies={currencies}
       currencyBalances={mmOrderBookTrade?.currencyBalances}
       isExpertMode={isExpertMode}
@@ -56,7 +81,8 @@ function MMCommitButtonComp({ mmOrderBookTrade, mmRFQTrade, mmQuoteExpiryRemaini
       wrapInputError={wrapInputError}
       recipient={recipient}
       onUserInput={onUserInput}
-      // isPendingError={isPendingError}
+      isPendingError={isPendingError}
+      currentAllowance={currentAllowance}
     />
   )
 }
