+++ apps/web/src/views/Swap/V3Swap/containers/MMCommitButtonV2.tsx
@@ -2,16 +2,15 @@ import { Currency } from '@pancakeswap/sdk'
 import { useExpertMode } from '@pancakeswap/utils/user'
 import { useCurrency } from 'hooks/Tokens'
 import { useIsTransactionUnsupported } from 'hooks/Trades'
-import { ApprovalState, useApproveCallback } from 'hooks/useApproveCallback'
 import useWrapCallback, { WrapType } from 'hooks/useWrapCallback'
-import { memo, useEffect, useMemo, useState } from 'react'
+import { memo, useMemo } from 'react'
 import { Field } from 'state/swap/actions'
 import { useSwapState } from 'state/swap/hooks'
 import { useSwapActionHandlers } from 'state/swap/useSwapActionHandlers'
-import { MMSwapCommitButtonV1 } from 'views/Swap/MMLinkPools/components/MMCommitButtonV1'
+import { MMSwapCommitButtonV2 } from 'views/Swap/MMLinkPools/components/MMCommitButtonV2'
 import { useAccount } from 'wagmi'
 
-function MMCommitButtonCompV1({ mmOrderBookTrade, mmRFQTrade, mmQuoteExpiryRemainingSec, mmTradeInfo }) {
+function MMCommitButtonCompV2({ mmOrderBookTrade, mmRFQTrade, mmQuoteExpiryRemainingSec, mmTradeInfo }) {
   const {
     typedValue,
     recipient,
@@ -19,14 +18,14 @@ function MMCommitButtonCompV1({ mmOrderBookTrade, mmRFQTrade, mmQuoteExpiryRemai
     [Field.OUTPUT]: { currencyId: outputCurrencyId },
   } = useSwapState()
 
-  const inputCurrency = useCurrency(inputCurrencyId)
-  const outputCurrency = useCurrency(outputCurrencyId)
+  const inputCurrency = useCurrency(inputCurrencyId) ?? undefined
+  const outputCurrency = useCurrency(outputCurrencyId) ?? undefined
   const { address: account } = useAccount()
   const [isExpertMode] = useExpertMode()
   const currencies: { [field in Field]?: Currency } = useMemo(
     () => ({
-      [Field.INPUT]: inputCurrency ?? undefined,
-      [Field.OUTPUT]: outputCurrency ?? undefined,
+      [Field.INPUT]: inputCurrency,
+      [Field.OUTPUT]: outputCurrency,
     }),
     [inputCurrency, outputCurrency],
   )
@@ -38,39 +37,15 @@ function MMCommitButtonCompV1({ mmOrderBookTrade, mmRFQTrade, mmQuoteExpiryRemai
     inputError: wrapInputError,
   } = useWrapCallback(inputCurrency, outputCurrency, typedValue)
   const showWrap = wrapType !== WrapType.NOT_APPLICABLE
-
-  const { approvalState, approveCallback, revokeCallback, currentAllowance, isPendingError } = useApproveCallback(
-    mmTradeInfo?.slippageAdjustedAmounts[Field.INPUT],
-    mmTradeInfo?.routerAddress,
-  )
-
-  // check if user has gone through approval process, used to show two step buttons, reset on token change
-  const [approvalSubmitted, setApprovalSubmitted] = useState<boolean>(false)
-
-  // Reset approval flow if input currency changed
-  useEffect(() => {
-    setApprovalSubmitted(false)
-  }, [inputCurrencyId])
-
-  // mark when a user has submitted an approval, reset onTokenSelection for input field
-  useEffect(() => {
-    if (approvalState === ApprovalState.PENDING) {
-      setApprovalSubmitted(true)
-    }
-  }, [approvalState, approvalSubmitted])
-
   const { onUserInput } = useSwapActionHandlers()
 
   return (
-    <MMSwapCommitButtonV1
+    <MMSwapCommitButtonV2
+      mmTradeInfo={mmTradeInfo}
       showWrap={showWrap}
-      approval={approvalState}
       swapIsUnsupported={swapIsUnsupported}
       account={account}
-      approvalSubmitted={approvalSubmitted}
       onWrap={onWrap}
-      approveCallback={approveCallback}
-      revokeCallback={revokeCallback}
       currencies={currencies}
       currencyBalances={mmOrderBookTrade?.currencyBalances}
       isExpertMode={isExpertMode}
@@ -81,10 +56,9 @@ function MMCommitButtonCompV1({ mmOrderBookTrade, mmRFQTrade, mmQuoteExpiryRemai
       wrapInputError={wrapInputError}
       recipient={recipient}
       onUserInput={onUserInput}
-      isPendingError={isPendingError}
-      currentAllowance={currentAllowance}
+      // isPendingError={isPendingError}
     />
   )
 }
 
-export const MMCommitButtonV1 = memo(MMCommitButtonCompV1)
+export const MMCommitButtonV2 = memo(MMCommitButtonCompV2)
