+++ apps/web/src/views/Swap/V3Swap/containers/index.ts
@@ -2,5 +2,5 @@ export * from './BuyCryptoLink'
 export * from './FormHeader'
 export * from './FormMain'
 export * from './PricingAndSlippage'
-export * from './SwapCommitButton'
+export * from './SwapCommitButtonV2'
 export * from './TradeDetails'
