+++ apps/web/src/views/Swap/V3Swap/hooks/index.ts
@@ -5,5 +5,5 @@ export * from './useRouterAddress'
 export * from './useSlippageAdjustedAmounts'
 export * from './useSwapBestTrade'
 export * from './useSwapCallback'
-export * from './useSwapCallbackV1'
+export * from './useSwapCallbackV2'
 export * from './useSwapInputError'
