+++ apps/web/src/views/Swap/V3Swap/hooks/useSwapCallArguments.ts
@@ -1,17 +1,13 @@
 /* eslint-disable @typescript-eslint/no-unused-vars */
 import { Percent, TradeType } from '@pancakeswap/sdk'
-import { SmartRouterTrade } from '@pancakeswap/smart-router/evm'
+import { SMART_ROUTER_ADDRESSES, SmartRouterTrade, SwapRouter } from '@pancakeswap/smart-router/evm'
 import { FeeOptions } from '@pancakeswap/v3-sdk'
 import { useMemo } from 'react'
-import {
-  PancakeSwapUniversalRouter,
-  getUniversalRouterAddress,
-  Permit2Signature,
-} from '@pancakeswap/universal-router-sdk'
-import { safeGetAddress } from 'utils'
-import useAccountActiveChain from 'hooks/useAccountActiveChain'
+
 import { useGetENSAddressByName } from 'hooks/useGetENSAddressByName'
-import { Address, Hex } from 'viem'
+
+import useAccountActiveChain from 'hooks/useAccountActiveChain'
+import { Address, Hex, isAddress } from 'viem'
 
 interface SwapCall {
   address: Address
@@ -31,7 +27,6 @@ export function useSwapCallArguments(
   trade: SmartRouterTrade<TradeType> | undefined | null,
   allowedSlippage: Percent,
   recipientAddressOrName: string | null | undefined,
-  permitSignature: Permit2Signature | undefined,
   deadline: bigint | undefined,
   feeOptions: FeeOptions | undefined,
 ): SwapCall[] {
@@ -40,9 +35,9 @@ export function useSwapCallArguments(
   const recipient = (
     recipientAddressOrName === null || recipientAddressOrName === undefined
       ? account
-      : safeGetAddress(recipientAddressOrName)
+      : isAddress(recipientAddressOrName)
       ? recipientAddressOrName
-      : safeGetAddress(recipientENSAddress)
+      : isAddress(recipientENSAddress ?? '')
       ? recipientENSAddress
       : null
   ) as Address | null
@@ -50,21 +45,71 @@ export function useSwapCallArguments(
   return useMemo(() => {
     if (!trade || !recipient || !account || !chainId) return []
 
-    const methodParameters = PancakeSwapUniversalRouter.swapERC20CallParameters(trade, {
+    const swapRouterAddress = chainId ? SMART_ROUTER_ADDRESSES[chainId] : undefined
+    if (!swapRouterAddress) return []
+
+    const { value, calldata } = SwapRouter.swapCallParameters(trade, {
       fee: feeOptions,
       recipient,
-      inputTokenPermit: permitSignature,
       slippageTolerance: allowedSlippage,
+      // ...(signatureData
+      //   ? {
+      //       inputTokenPermit:
+      //         'allowed' in signatureData
+      //           ? {
+      //               expiry: signatureData.deadline,
+      //               nonce: signatureData.nonce,
+      //               s: signatureData.s,
+      //               r: signatureData.r,
+      //               v: signatureData.v as any,
+      //             }
+      //           : {
+      //               deadline: signatureData.deadline,
+      //               amount: signatureData.amount,
+      //               s: signatureData.s,
+      //               r: signatureData.r,
+      //               v: signatureData.v as any,
+      //             },
+      //     }
+      //   : {}),
+
       deadlineOrPreviousBlockhash: deadline?.toString(),
     })
-    const swapRouterAddress = getUniversalRouterAddress(chainId)
-    if (!swapRouterAddress) return []
+
+    // if (argentWalletContract && trade.inputAmount.currency.isToken) {
+    //   return [
+    //     {
+    //       address: argentWalletContract.address,
+    //       calldata: argentWalletContract.interface.encodeFunctionData('wc_multiCall', [
+    //         [
+    //           approveAmountCalldata(trade.maximumAmountIn(allowedSlippage), swapRouterAddress),
+    //           {
+    //             to: swapRouterAddress,
+    //             value,
+    //             data: calldata,
+    //           },
+    //         ],
+    //       ]),
+    //       value: '0x0',
+    //     },
+    //   ]
+    // }
     return [
       {
         address: swapRouterAddress,
-        calldata: methodParameters.calldata as `0x${string}`,
-        value: methodParameters.value as `0x${string}`,
+        calldata,
+        value,
       },
     ]
-  }, [account, allowedSlippage, chainId, deadline, feeOptions, recipient, permitSignature, trade])
+  }, [
+    account,
+    allowedSlippage,
+    // argentWalletContract,
+    chainId,
+    deadline,
+    feeOptions,
+    recipient,
+    // signatureData,
+    trade,
+  ])
 }
