+++ apps/web/src/views/Swap/V3Swap/hooks/useSwapCallback.ts
@@ -2,9 +2,8 @@
 import { useTranslation } from '@pancakeswap/localization'
 import { TradeType } from '@pancakeswap/sdk'
 import { SmartRouterTrade } from '@pancakeswap/smart-router/evm'
-import { Permit2Signature } from '@pancakeswap/universal-router-sdk'
 import { FeeOptions } from '@pancakeswap/v3-sdk'
-import { useMemo } from 'react'
+import { ReactNode, useMemo } from 'react'
 
 import { useUserSlippage } from '@pancakeswap/utils/user'
 import { INITIAL_ALLOWED_SLIPPAGE } from 'config/constants'
@@ -14,10 +13,12 @@ import { basisPointsToPercent } from 'utils/exchange'
 import useAccountActiveChain from 'hooks/useAccountActiveChain'
 import { SendTransactionResult } from 'wagmi/actions'
 import useSendSwapTransaction from './useSendSwapTransaction'
-import { useSwapCallArguments } from './useSwapCallArguments'
-import type { TWallchainMasterInput, WallchainStatus } from './useWallchain'
 
-export enum SwapCallbackState {
+import { useSwapCallArguments } from './useSwapCallArguments'
+import type { TWallchainMasterInput } from './useWallchain'
+import { useWallchainSwapCallArguments } from './useWallchain'
+
+enum SwapCallbackState {
   INVALID,
   LOADING,
   VALID,
@@ -27,18 +28,17 @@ export enum SwapCallbackState {
 interface UseSwapCallbackReturns {
   state: SwapCallbackState
   callback?: () => Promise<SendTransactionResult>
-  error?: string
+  error?: ReactNode
   reason?: string
 }
 interface UseSwapCallbackArgs {
   trade: SmartRouterTrade<TradeType> | undefined | null // trade to execute, required
   // allowedSlippage: Percent // in bips
   // recipientAddressOrName: string | null | undefined // the ENS name or address of the recipient of the trade, or null if swap should be returned to sender
+  // signatureData: SignatureData | null | undefined
   deadline?: bigint
-  permitSignature: Permit2Signature | undefined
   feeOptions?: FeeOptions
-  onWallchainDrop?: () => void
-  statusWallchain?: WallchainStatus
+  onWallchainDrop: () => void
   wallchainMasterInput?: TWallchainMasterInput
 }
 
@@ -46,9 +46,11 @@ interface UseSwapCallbackArgs {
 // and the user has approved the slippage adjusted input amount for the trade
 export function useSwapCallback({
   trade,
+  // signatureData,
   deadline,
-  permitSignature,
   feeOptions,
+  onWallchainDrop,
+  wallchainMasterInput,
 }: UseSwapCallbackArgs): UseSwapCallbackReturns {
   const { t } = useTranslation()
   const { account, chainId } = useAccountActiveChain()
@@ -61,21 +63,26 @@ export function useSwapCallback({
     trade,
     allowedSlippage,
     recipientAddress,
-    permitSignature,
+    // signatureData,
     deadline,
     feeOptions,
   )
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  // const wallchainSwapCalls = useWallchainSwapCallArguments(
-  //   trade,
-  //   swapCalls,
-  //   account,
-  //   onWallchainDrop,
-  //   wallchainMasterInput,
-  // )
-  // const wallchainSwapCalls = []
+  const wallchainSwapCalls = useWallchainSwapCallArguments(
+    trade,
+    swapCalls,
+    account,
+    onWallchainDrop,
+    wallchainMasterInput,
+  )
 
-  const { callback } = useSendSwapTransaction(account, chainId, trade ?? undefined, swapCalls, 'UniversalRouter')
+  const { callback } = useSendSwapTransaction(
+    account,
+    chainId,
+    trade,
+    // @ts-expect-error uncompatible types side-by-side cause wrong type assertion
+    wallchainSwapCalls,
+    'V3SwapRouter',
+  )
 
   return useMemo(() => {
     if (!trade || !account || !chainId || !callback) {
